一旦リファクタリング
