+++ goodcareer-functions/functions/index.js
@@ -1,158 +1,20 @@
 const functions = require('firebase-functions');
-const admin = require('firebase-admin');
+
 const app = require('express')();
 
-admin.initializeApp();
+const FBAuth = require('./util/fbAuth');
 
-const config = {
-  apiKey: "AIzaSyBCak4ASMj3VIQkDVbIQ9Wkuz4gbVFqV7c",
-  authDomain: "goodcareer-746e4.firebaseapp.com",
-  databaseURL: "https://goodcareer-746e4.firebaseio.com",
-  projectId: "goodcareer-746e4",
-  storageBucket: "goodcareer-746e4.appspot.com",
-  messagingSenderId: "53080588448",
-  appId: "1:53080588448:web:2d03c1d362c832bc7c49c5",
-  measurementId: "G-LS4ZYRH9C8"
-};
+const { getAllCompanies, postOneCompany } = require('./handlers/companies')
+const { signup, login } = require('./handlers/users')
 
 
-const firebase = require('firebase');
-firebase.initializeApp(config);
-
-const db = admin.firestore();
-
-app.get('/companies', (req, res) => {
-  db.collection('companies').orderBy('createdAt', 'desc').get()
-    .then(data => {
-      let companies = [];
-      data.forEach(doc => {
-        companies.push({
-          companyId: doc.id,
-          ...doc.data()
-        });
-      });
-      return res.json(companies);
-    })
-    .catch(err => console.error(err));
-})
-
-app.post('/company', (req, res) => {
-  const newCompany = {
-    body: req.body.body,
-    industry: req.body.industry,
-    createdAt: new Date().toISOString()
-  };
-
-  db
-    .collection('companies')
-    .add(newCompany)
-    .then(doc => {
-      res.json({ message: `document ${doc.id} created successfully`});
-    })
-    .catch(err => {
-      res.status(500).json({ error: 'something went wrong'});
-      console.error(err);
-    });
-});
-
-const isEmail = (email) => {
-  const regEx = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-  if(email.match(regEx)) return true;
-  else return false;
-}
-
-const isEmpty = (string) => {
-  if(string.trim() === '') return true;
-  else return false;
-}
+app.get('/companies', getAllCompanies)
+app.post('/company', FBAuth, postOneCompany);
 
 // SIGNUP route
-app.post('/signup', (req, res) => {
-  const newUser = {
-    email: req.body.email,
-    password: req.body.password,
-    confirmPassword: req.body.confirmPassword,
-    handle: req.body.handle
-  };
-
-  let errors = {};
-
-  if(isEmpty(newUser.email)) {
-    errors.email = 'Email must not be empty'
-  } else if (!isEmail(newUser.email)) {
-    errors.email = 'Must be a valid email address'
-  }
-
-  if(isEmpty(newUser.password)) errors.password = 'Must not be empty';
-  if(newUser.password !== newUser.confirmPassword) errors.confirmPassword = 'Passwords must match';
-  if(isEmpty(newUser.handle)) errors.handle = 'Must not be empty';
-
-  if(Object.keys(errors).length > 0) return res.status(400).json(errors)
-
-  let token, userId;
-  db.doc(`/users/${newUser.handle}`).get()
-    .then(doc => {
-      if(doc.exists){
-        return res.status(400).json({ handle: 'this handle is already taken'});
-      } else {
-        return firebase
-      .auth().createUserWithEmailAndPassword(newUser.email, newUser.password);
-      }
-    })
-    .then(data => {
-      userId = data.user.uid;
-      return data.user.getIdToken();
-    })
-    .then(idToken => {
-      token = idToken;
-      const userCredentials = {
-        handle: newUser.handle,
-        email: newUser.email,
-        createdAt: new Date().toISOString(),
-        userId
-      };
-      return db.doc(`/users/${newUser.handle}`).set(userCredentials);
-    })
-    .then(() => {
-      return res.status(201).json({ token });
-    })
-    .catch(err => {
-      if(err.code === 'auth/email-already-in-use'){
-        return res.status(400).json({email: 'Email is already in use'})
-      } else {
-        return res.status(500).json({ error: err.code })
-      }
-    })
-});
+app.post('/signup', signup);
 
 // Login
-app.post('/login', (req, res) => {
-  const user = {
-    email: req.body.email,
-    password: req.body.password
-  };
-
-  let errors = {};
-
-  if(isEmpty(user.email)) errors.email = 'Must not be empty';
-  if(isEmpty(user.password)) errors.password = 'Must not be empty';
-
-  if(Object.keys(errors).length > 0) return res.status(400).json(errors);
-
-  firebase.auth().signInWithEmailAndPassword(user.email, user.password)
-    .then(data => {
-      return data.user.getIdToken();
-    })
-    .then(token => {
-      return res.json({token});
-    })
-    .catch(err => {
-      console.error(err);
-      if (err.code === 'auth/wrong-password') {
-        return res.status(403).json({ general: 'Wrong credentials, please try again'});
-      }
-       else return res.status(500).json({error: err.code});
-    })
-})
+app.post('/login', login);
 
 exports.api = functions.region('asia-northeast1').https.onRequest(app);
