+++ goodcareer-functions/functions/util/config.js
@@ -1,0 +1,10 @@
+module.exports = {
+  apiKey: "AIzaSyBCak4ASMj3VIQkDVbIQ9Wkuz4gbVFqV7c",
+  authDomain: "goodcareer-746e4.firebaseapp.com",
+  databaseURL: "https://goodcareer-746e4.firebaseio.com",
+  projectId: "goodcareer-746e4",
+  storageBucket: "goodcareer-746e4.appspot.com",
+  messagingSenderId: "53080588448",
+  appId: "1:53080588448:web:2d03c1d362c832bc7c49c5",
+  measurementId: "G-LS4ZYRH9C8"
+};
